requirements.txt
grpcio
opencv-python
numpy
tensorflow-serving-api
